refactor: update issues components imports to use shadcn directory
