+++ features/issues/components/issue-attachments-uploader.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import { useRef } from 'react';
-import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
+import { Button } from '@/components/shadcn/button';
+import { Input } from '@/components/shadcn/input';
 import { Upload, Loader2 } from 'lucide-react';
 import { toast } from '@/lib/styles/toast-styles';
 import { useUpdateIssueWithFiles } from '@/hooks/issue';
